+++ testcanarybot/__init__.py
@@ -2,4 +2,4 @@ from .source.application import app
 from .source.library import init_async
 
 __all__ = ['enums', 'exceptions', 'objects', 'tools']
-__version__ = "00.09.101 dev"
+__version__ = "00.09.102 dev"
